refactor(quiz): migrate quiz feature to app router and restructure architecture
- Move quiz pages from pages/[language]/quiz to app/[language]/(guest)/(empty-layout)/quiz
- Restructure quiz feature from features/quiz to entities/quiz and widgets/quiz
- Extract quiz models, types, and business logic to entities layer
- Move quiz UI components to widgets layer
- Add new quiz layout with font preloading and QuizProvider
- Implement static generation for quiz question and result pages
- Add image capture utility for result sharing functionality
- Update imports and dependencies throughout the codebase

This refactoring improves code organization by following domain-driven design
principles and prepares the quiz feature for better maintainability and
scalability in the app router architecture.
