+++ app/[language]/(guest)/(empty-layout)/quiz/result/page.tsx
@@ -1,0 +1,10 @@
+import { Metadata } from 'next';
+import { QuizResultWidget } from '@/widgets/quiz';
+
+export const metadata: Metadata = {
+  title: '測驗結果',
+};
+
+export default function QuizResultPage() {
+  return <QuizResultWidget />;
+}
