+++ entities/quiz/index.ts
@@ -1,3 +1,3 @@
 export * from './model';
 export * from './ui';
-export * from './lib';
+export * from './api';
