+++ entities/quiz/model/index.ts
@@ -1,0 +1,5 @@
+export * from './types';
+export * from './store';
+export * from './question-map';
+export * from './result-detail-map';
+export * from './theme-map';
