--- features/quiz/components/index.ts
@@ -1,7 +1,0 @@
-import dynamic from 'next/dynamic';
-
-export * from './Styled';
-
-export const ResultChart = dynamic(() => import('./ResultChart'), {
-  ssr: false,
-});
