--- features/quiz/constants/index.ts
@@ -1,3 +1,0 @@
-export * from './questionMap';
-export * from './resultDetailMap';
-export * from './themeMap';
