--- features/quiz/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useResultStyles } from './useResultStyles';
