+++ features/quiz/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './use-result-styles';
