+++ features/quiz/ui/styled.tsx
