+++ widgets/quiz/ui/index.ts
@@ -1,0 +1,4 @@
+export * from './quiz-intro-widget';
+export * from './quiz-question-widget';
+export * from './quiz-result-widget';
+export * from './quiz-result-detail-widget';
